Unified all exceptions classes
