+++ resources/lib/kodi/library_utils.py
@@ -17,10 +17,11 @@ from functools import wraps
 import xbmc
 
 from resources.lib import common
-from resources.lib.utils.api_paths import PATH_REQUEST_SIZE_STD
+from resources.lib.common.exceptions import InvalidVideoId
 from resources.lib.database.db_utils import VidLibProp
 from resources.lib.globals import G
 from resources.lib.kodi import nfo, ui
+from resources.lib.utils.api_paths import PATH_REQUEST_SIZE_STD
 from resources.lib.utils.logging import LOG, measure_exec_time_decorator
 
 LIBRARY_HOME = 'library'
@@ -76,7 +77,7 @@ def is_videoid_in_db(videoid):
         return G.SHARED_DB.episode_id_exists(videoid.tvshowid,
                                              videoid.seasonid,
                                              videoid.episodeid)
-    raise common.InvalidVideoId('videoid {} type not implemented'.format(videoid))
+    raise InvalidVideoId('videoid {} type not implemented'.format(videoid))
 
 
 def get_episode_title_from_path(file_path):
