+++ resources/lib/services/cache/cache_management.py
@@ -15,10 +15,10 @@ from datetime import datetime, timedelta
 from functools import wraps
 from time import time
 
-from resources.lib.utils.exceptions import UnknownCacheBucketError, CacheMiss
 from resources.lib.common import G
-from resources.lib.database.db_exceptions import SQLiteConnectionError, SQLiteError, ProfilesMissing
 from resources.lib.common.cache_utils import BUCKET_NAMES, BUCKETS
+from resources.lib.common.exceptions import (UnknownCacheBucketError, CacheMiss, DBSQLiteConnectionError,
+                                             DBSQLiteError, DBProfilesMissing)
 from resources.lib.utils.logging import LOG
 
 CONN_ISOLATION_LEVEL = None  # Autocommit mode
@@ -38,7 +38,7 @@ def handle_connection(func):
             return func(*args, **kwargs)
         except sql.Error as exc:
             LOG.error('SQLite error {}:', exc.args[0])
-            raise SQLiteConnectionError
+            raise DBSQLiteConnectionError
         finally:
             if conn:
                 args[0].is_connected = False
@@ -133,7 +133,7 @@ class CacheManagement(object):
             if bucket['is_persistent']:
                 return self._get_db(bucket['name'], identifier)
             raise CacheMiss()
-        except ProfilesMissing:
+        except DBProfilesMissing:
             # Raised by _add_prefix there is no active profile guid when add-on is installed from scratch
             raise CacheMiss()
 
@@ -152,7 +152,7 @@ class CacheManagement(object):
             return result[0]
         except sql.Error as exc:
             LOG.error('SQLite error {}:', exc.args[0])
-            raise SQLiteError
+            raise DBSQLiteError
 
     def add(self, bucket, identifier, data, ttl=None, expires=None):
         """
@@ -176,7 +176,7 @@ class CacheManagement(object):
             if bucket['is_persistent']:
                 # Save the item data to the cache database
                 self._add_db(bucket['name'], identifier, data, expires)
-        except ProfilesMissing:
+        except DBProfilesMissing:
             # Raised by _add_prefix there is no active profile guid when add-on is installed from scratch
             pass
 
@@ -189,7 +189,7 @@ class CacheManagement(object):
             cursor.execute(query, (bucket_name, identifier, sql.Binary(data), expires, int(time())))
         except sql.Error as exc:
             LOG.error('SQLite error {}:', exc.args[0])
-            raise SQLiteError
+            raise DBSQLiteError
 
     def delete(self, bucket, identifier, including_suffixes):
         """
@@ -214,7 +214,7 @@ class CacheManagement(object):
             if bucket['is_persistent']:
                 # Delete the item data from cache database
                 self._delete_db(bucket['name'], identifier, including_suffixes)
-        except ProfilesMissing:
+        except DBProfilesMissing:
             # Raised by _add_prefix there is no active profile guid when add-on is installed from scratch
             pass
 
@@ -230,7 +230,7 @@ class CacheManagement(object):
             cursor.execute(query, (bucket_name, identifier))
         except sql.Error as exc:
             LOG.error('SQLite error {}:', exc.args[0])
-            raise SQLiteError
+            raise DBSQLiteError
 
     def clear(self, buckets=None, clear_database=True):
         """
@@ -265,7 +265,7 @@ class CacheManagement(object):
                 cursor.execute(query, (bucket['name'], ))
         except sql.Error as exc:
             LOG.error('SQLite error {}:', exc.args[0])
-            raise SQLiteError
+            raise DBSQLiteError
 
     def delete_expired(self):
         bucket_names_db = []
@@ -294,7 +294,7 @@ class CacheManagement(object):
             cursor.execute(query, bucket_names)
         except sql.Error as exc:
             LOG.error('SQLite error {}:', exc.args[0])
-            raise SQLiteError
+            raise DBSQLiteError
 
 
 def _compute_next_schedule():
