+++ resources/lib/services/msl/default_crypto.py
@@ -13,6 +13,7 @@ from __future__ import absolute_import, division, unicode_literals
 import base64
 import json
 
+
 try:  # The crypto package depends on the library installed (see Wiki)
     from Crypto.Random import get_random_bytes
     from Crypto.Hash import HMAC, SHA256
@@ -28,9 +29,9 @@ except ImportError:
     from Cryptodome.Util import Padding
     from Cryptodome.Cipher import AES
 
+from resources.lib.common.exceptions import MSLError
 from resources.lib.utils.logging import LOG
 from .base_crypto import MSLBaseCrypto
-from .exceptions import MSLError
 
 
 class DefaultMSLCrypto(MSLBaseCrypto):
