--- resources/lib/services/msl/exceptions.py
@@ -1,22 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    Copyright (C) 2017 Sebastian Golasch (plugin.video.netflix)
-    Copyright (C) 2018 Caphm (original implementation module)
-    Common MSL exceptions
-
-    SPDX-License-Identifier: MIT
-    See LICENSES/MIT.md for more information.
-"""
-from __future__ import absolute_import, division, unicode_literals
-
-
-class MSLError(Exception):
-    pass
-
-
-class LicenseError(MSLError):
-    pass
-
-
-class ManifestError(MSLError):
-    pass
