+++ resources/lib/services/nfsession/directorybuilder/dir_builder.py
@@ -12,7 +12,7 @@ from __future__ import absolute_import, division, unicode_literals
 from future.utils import iteritems
 
 from resources.lib.utils.data_types import merge_data_type
-from resources.lib.utils.exceptions import CacheMiss
+from resources.lib.common.exceptions import CacheMiss
 from resources.lib.common import VideoId
 from resources.lib.globals import G
 from resources.lib.services.nfsession.directorybuilder.dir_builder_items \
